Add intel-conf methods to Darwin
There has existed for awhile a patch to pciutils for macOS/Darwin that uses DirectHW.kext (service named DirectHWService) to support intel-conf access methods. This change incorporates those patches for the following reasons:
1) To expose the intel-conf methods for macOS/Darwin.
2) macOS 10.11 El Capitan does not support the darwin access method (service named AppleACPIPlatformExpert is missing).
3) Loading DirectHW.kext does not require a restart while AppleACPIPlatformExpert may require a restart because it depends on certain kernel debug flags being set.

The source I used for the DirectHW patches is from coreboot. I've included DirectHW.c and DirectHW.h so that the DirectHW framework is not required for building or running.