Add Xcode projects
For easier development and debugging on macOS for macOS.

pciutils10.6.xcodeproj can be used for building/debugging ppc, ppc64, i386, x86_64 on Mac OS X 10.6. The resulting binaries work with Mac OS X 10.4 to macOS 13.
pciutils.xcodeproj can be used for building/debugging x86_64 and arm64. arm64 is not tested.
The recommended method for building is to use the "make clean", "make", "sudo make install" commands.