Add darwin3 access method
This is similar to the darwin2 access method except we use the DirectHW IOUserClient instead of the IOPCIBridge IOUserClient to access config registers. The segment number is used to select the root IOPCIBridge.
This access method should work with all Mac architectures (arm64e is not tested).
This is the only access method that works for PowerPC Macs.
This may be the only access method to allow PCIe extended config space access for OS X 10.11 or 10.12 since those macOS versions seem to have issues with the darwin (AppleACPIPlatformExpert) and darwin2 (IOPCIBridge) access methods and the intel-conf1 access method is limited to PCI config space registers.
This access method requires the DirectHW.kext from my fork of DirectHW to be loaded.